+++ packages/memo/test/writable.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect } from "vitest";
 import { createWritableMemo } from "../src";
-import { createRoot, createSignal } from "solid-js";
+import { batch, createRoot, createSignal } from "solid-js";
 
 describe("createWritableMemo", () => {
   test("behaves like a memo", () =>
@@ -27,4 +27,51 @@ describe("createWritableMemo", () => {
       expect(result()).toBe(3);
       dispose();
     }));
+
+  test("consistency of previous value in the callbacks", () => {
+    const [count, setCount] = createSignal(1);
+    let prevCb: number | undefined;
+    const { result, setResult, dispose } = createRoot(dispose => {
+      const [result, setResult] = createWritableMemo(p => ((prevCb = p), count() * 2), -2);
+      expect(prevCb).toBe(-2);
+      expect(result()).toBe(2);
+      return { result, dispose, setResult };
+    });
+    setResult(p => {
+      expect(p).toBe(2);
+      return 5;
+    });
+    expect(result()).toBe(5);
+    expect(prevCb).toBe(-2);
+    setCount(5);
+    expect(result()).toBe(count() * 2);
+    expect(prevCb).toBe(5);
+    setCount(7);
+    expect(prevCb).toBe(10);
+    setResult(p => {
+      expect(p).toBe(14);
+      return 3;
+    });
+    expect(result()).toBe(3);
+    dispose();
+  });
+
+  test("updating and reading in a batch", () => {
+    createRoot(dispose => {
+      const [source, setSource] = createSignal(1);
+      const [memo, setMemo] = createWritableMemo(source, -2);
+
+      batch(() => {
+        setSource(2);
+        expect(memo()).toBe(2);
+      });
+
+      batch(() => {
+        setMemo(-3);
+        expect(memo()).toBe(-3);
+      });
+
+      dispose();
+    });
+  });
 });