+++ packages/clipboard/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/clipboard
 
+## 1.5.2-beta.0
+
+### Patch Changes
+
+- Updated dependencies [d6559a32]
+  - @solid-primitives/utils@5.4.0-beta.0
+
 ## 1.5.1
 
 ### Patch Changes