+++ packages/jsx-parser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/jsx-parser
 
+## 0.1.3-beta.0
+
+### Patch Changes
+
+- Updated dependencies [d6559a32]
+  - @solid-primitives/utils@5.4.0-beta.0
+
 ## 0.1.2
 
 ### Patch Changes