+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @solid-primitives/styles
 
+## 0.0.105-beta.0
+
+### Patch Changes
+
+- Updated dependencies [d6559a32]
+  - @solid-primitives/utils@5.4.0-beta.0
+  - @solid-primitives/rootless@1.2.6-beta.0
+
 ## 0.0.104
 
 ### Patch Changes